test: fold relative-date metric filter tests into existing suites
Move the PROD-8219 regression tests into the conventional co-located suites
and delete the standalone prod8219 files:
- query-time swap cases -> MetricQueryBuilder.test.ts
- compile-time capture cases -> exploreCompiler.test.ts
